Rebase onto master, clean up some old invocations of the swarm_train and moving the dict stuff into utils as opposed to core, since they are utils.
This necessitates the move of the tests too
